🐛 Modify : 교환일기 load-diary 페이지에 부분 로딩 UI 추가 및 refetch 로직 제거 #203
